requirements.txt
Django==2.2.16
django_autocomplete_light==3.8.1
PyMySQL==0.9.3
django_crispy_forms==1.9.2